Fix, incorrect use of std::forward Use pascal case enum values. Use int instead of unsigned int where possible. Use typename instead of class in template.

+++ src/net/lanudpsender.cpp
@@ -4,7 +4,7 @@ using namespace std::chrono_literals;
 
 namespace net {
 
-	LanUdpSender::LanUdpSender(asio::io_context& ioContext, size_t maxSize)
+	LanUdpSender::LanUdpSender(asio::io_context& ioContext, int maxSize)
 		: ioContext_{ioContext}
 		, socket_{ioContext}
 		, protobufMessage_{maxSize}
